frontend/pyproject.toml
[project]
name = "pipecat-quickstart"
version = "0.1.0"
description = "Quickstart example for building voice AI bots with Pipecat"
requires-python = ">=3.10"
dependencies = [
    "pipecat-ai[webrtc,daily,silero,deepgram,openai,cartesia,runner]>=0.0.83",
    "pipecatcloud>=0.2.4",
    "httpx>=0.27.0",
    "fastapi>=0.112.0",
    "uvicorn>=0.30.5",
    "python-dotenv>=1.0.1",
    "pipecat>=0.3.0",
]

[dependency-groups]
dev = [
    "ruff~=0.12.1",
]

[tool.ruff]
line-length = 100
[tool.ruff.lint]
select = ["I"]
